Consolidate common ryu c functions
